requirements.txt
torch==2.0.1
transformers==4.37.1
diffusers==0.18.0
protobuf==3.20.3
pyarrow==14.0.1
langchain==0.1.3
sentence-transformers==2.2.2
python-dotenv==1.0.1
more-itertools==10.2.0
pinecone-client[grpc]==3.0.2
pandas==2.0.0
openai==1.12.0
canopy-sdk==0.8.0